Can now save new parts and modify parts.
Now have a functioning search bar, case-insensitive.

Search functions now match UML diagram correctly.

One more save before exception handling.

Added Product Tables and Screen with Buttons.  Both searches working for main now.

All buttons are working now.  There are some logic glitches in passing data and displaying tableviews.

Looking done.  Just adding JavaDoc comments and DRYing out some code.

Requirements Fulfilled.

+++ src/controller/MainController.java
@@ -1,7 +1,3 @@
-/**
- * @author Heaven-Leigh (Michelle) Masters
- */
-
 package controller;
 
 import javafx.scene.control.*;
@@ -26,7 +22,12 @@ import java.net.URL;
 import java.util.Optional;
 import java.util.ResourceBundle;
 
-
+/**
+ * controller for the main screen of the application.
+ * Shows table of parts and products in Inventory.
+ * Allows searching and deleting of both parts and products.
+ * @author Heaven-Leigh (Michelle) Masters
+ */
 public class MainController implements Initializable {
 
     @FXML
@@ -80,9 +81,9 @@ public class MainController implements Initializable {
     /**
      * Called to initialize a controller after its root element has been completely processed,
      *                  and sets up the tables for parts and products.
-     * @param location The location used to resolve relative paths for the root object,
-     *                 or null if the location is not known.
-     * @param resources The resources used to localize the root object, or null if the root object was not localized.
+     * @param location used to resolve relative paths for the root object,
+     *                 or null if the location is not known
+     * @param resources used to localize the root object, or null if the root object was not localized
      */
     @Override
     public void initialize(URL location, ResourceBundle resources) {
@@ -106,9 +107,8 @@ public class MainController implements Initializable {
 
     /**
      * Exits/Closes the application.
-     * @param actionEvent triggered by pressing the exit button.
+     * @param actionEvent triggered by pressing the exit button
      */
-
     public void OnExitProgramButton(ActionEvent actionEvent) {
 
         Stage stage = (Stage) OnExitProgramButton.getScene().getWindow();
@@ -119,26 +119,15 @@ public class MainController implements Initializable {
 /** Parts (Left side) *************************************************************************/
 
     /**
-     * Searches for a substring of a part name or ID to locate any matching parts.
-     * @param keyEvent triggered from key entered in search text field
+     * searches the search string entered to look for any matches to part ID or part name.
+     * Uses FindMatchedParts method to search for any matching parts.
+     * Fills a matchedParts list with parts that match the search, and sets the parts table with them.
+     * @param keyEvent triggered by a key press within the part search text field
      */
-
     public void OnSearchStringEntered(KeyEvent keyEvent) {
 
         String searchString = OnSearchText.getText();
-        boolean isNumericString = isNumeric(searchString);
-        ObservableList<Part> matchedParts = FXCollections.observableArrayList();
-
-        if (isNumericString) {
-
-            matchedParts.add(Inventory.lookupPart(Integer.parseInt(searchString)));
-        }
-
-        else {
-
-            matchedParts = Inventory.lookupPart(OnSearchText.getText());
-        }
-
+        ObservableList<Part> matchedParts = FindMatchedParts(searchString);
         partsTableView.setItems(matchedParts);
     }
 
@@ -147,7 +136,6 @@ public class MainController implements Initializable {
      * @param numOrString receives a String to check whether it is numeric.
      * @return true if the String is numeric, otherwise false.
      */
-
     public static boolean isNumeric(String numOrString) {
 
         if (numOrString == null) {
@@ -167,7 +155,7 @@ public class MainController implements Initializable {
     }
 
     /**
-     * This method opens the scene for adding a new part.
+     * Opens the scene for adding a new part.
      *
      * @param actionEvent triggered from pressing the part's add button.
      * @throws IOException triggered if the scene cannot be loaded.
@@ -185,9 +173,8 @@ public class MainController implements Initializable {
 
     /**
      * Passes a selected part to the modify part screen for modification.
-     *
+     * Catches the exception and shows error dialog if a part is not selected.
      * @param actionEvent triggered by pressing the part's modify button
-     * @throws Exception caught when a part is not selected
      */
     public void OnPartModifyButton(ActionEvent actionEvent) {
         System.out.println("Pressed Part Modify Button");
@@ -215,8 +202,8 @@ public class MainController implements Initializable {
 
     /**
      * This method will delete a selected part from inventory upon confirmation.
-     * @param actionEvent triggered from the part's delete button.
-     * @exception Exception triggered when a part is not selected.
+     * Catches the exception and shows an error dialog if a part is not selected.
+     * @param actionEvent triggered from the part's delete button
      */
     public void OnPartDeleteButton(ActionEvent actionEvent) {
         try {
@@ -241,9 +228,8 @@ public class MainController implements Initializable {
     /**
      * Changes the scene for the addition of a new product.
      * @param actionEvent triggered from pressing the product's add button.
-     * @throws IOException triggered if the scene cannot be loaded.
+     * @throws IOException if the scene cannot be loaded.
      */
-
     public void OnProductAddButton(ActionEvent actionEvent) throws IOException {
         FXMLLoader loader = new FXMLLoader();
         loader.setLocation(getClass().getResource("../view/AddModifyProductScene.fxml"));
@@ -264,9 +250,10 @@ public class MainController implements Initializable {
 
     /**
      * Searches for a substring of a product name or ID to locate any matching parts.
+     * Uses same strategy as searching for a matchedParts,
+     * but with lookupProduct instead of lookupPart from Inventory.
      * @param keyEvent triggered from key entered in search text field
      */
-
     public void OnSearchStringEnteredProduct(KeyEvent keyEvent) {
         String searchString = OnSearchTextProduct.getText();
         boolean isNumericString = isNumeric(searchString);
@@ -277,12 +264,18 @@ public class MainController implements Initializable {
             matchedProducts.add(Inventory.lookupProduct(Integer.parseInt(searchString)));
         } else {
 
-            matchedProducts = Inventory.lookupProduct(OnSearchTextProduct.getText());
+            matchedProducts = Inventory.lookupProduct(searchString);
         }
 
         productsTableView.setItems(matchedProducts);
     }
 
+    /**
+     * changes the scene to the modify product screen.
+     * Calls to initialize the selected products information on the modify screen.
+     * Catches the error that a product was not selected if one is not selected.
+     * @param actionEvent triggered by pressing the modify button for products
+     */
     public void OnProductModifyButton(ActionEvent actionEvent) {
         System.out.println("Pressed Product Modify Button");
         try {
@@ -295,7 +288,7 @@ public class MainController implements Initializable {
             AddModifyProductController controller = loader.getController();
             controller.AddModifyProductLabel.setText("Modify Product");
             controller.InitializePickPartsTable();
-            controller.InitProduct((Product) productsTableView.getSelectionModel().getSelectedItem());
+            controller.InitProduct(productsTableView.getSelectionModel().getSelectedItem());
 
             Stage stage = (Stage) ((Button) actionEvent.getSource()).getScene().getWindow();
             stage.setTitle("Modify Part");
@@ -309,12 +302,22 @@ public class MainController implements Initializable {
         }
     }
 
-
-
+    /**
+     * deletes a product from the inventory upon confirmation.
+     * Checks to make sure the product has no parts in it, otherwise throws error.
+     * @param actionEvent triggered from delete button under product's table on main screen
+     */
     public void OnProductDeleteButton(ActionEvent actionEvent) {
+
         try {
 
-            Product selectedProduct = (Product) productsTableView.getSelectionModel().getSelectedItem();
+            Product selectedProduct = productsTableView.getSelectionModel().getSelectedItem();
+
+            if (!selectedProduct.getAllAssociatedParts().isEmpty()) {
+                ErrorException("Cannot Delete Product", "Remove all parts before " +
+                        selectedProduct.getName() + " can be deleted");
+                return;
+            }
 
             if (DeleteConfirmation("Confirm you want to delete" +
                     selectedProduct.getName())) {
@@ -324,12 +327,43 @@ public class MainController implements Initializable {
 
         } catch (Exception e) {
 
-            ErrorException("Part not found",
+            ErrorException("Product not found",
                     "Please select a product in order to delete a product.");
         }
     }
 
-    public boolean DeleteConfirmation(String contentText) {
+    /**
+     * finds matched parts given a search string.
+     * First checks if it is numeric and if so matches the part ID number.
+     * Else checks for matches within the part name.
+     * @param searchString uses the search text that has been entered into the field
+     * @return any matched parts, or empty list which causes all parts to display
+     */
+    public static ObservableList<Part> FindMatchedParts(String searchString) {
+
+        boolean isNumericString = isNumeric(searchString);
+        ObservableList<Part> matchedParts = FXCollections.observableArrayList();
+
+        if (isNumericString) {
+
+            matchedParts.add(Inventory.lookupPart(Integer.parseInt(searchString)));
+        }
+
+        else {
+
+            matchedParts = Inventory.lookupPart(searchString);
+        }
+
+        return matchedParts;
+    }
+
+    /**
+     * displays a delete confirmation dialog box and displays a method's context text.
+     * @param contentText is the text from the method with more information regarding details of delete
+     * @return boolean true if the okay button is pressed for deleting, and false if cancelled.
+     */
+    public static boolean DeleteConfirmation(String contentText) {
+
         Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
         alert.setTitle("Confirmation");
         alert.setHeaderText("Delete item?");
@@ -343,12 +377,17 @@ public class MainController implements Initializable {
         return false;
     }
 
-    public void ErrorException (String headerText, String contentText) {
+    /**
+     * displays an error dialog box with the header text and content text details from calling method.
+     * @param headerText is the heading or category of the error
+     * @param contentText is the details passed from calling method regarding the error.
+     */
+    public static void ErrorException (String headerText, String contentText) {
+
         Alert alert = new Alert(Alert.AlertType.ERROR);
         alert.setTitle("Error");
         alert.setHeaderText(headerText);
         alert.setContentText(contentText);
         alert.showAndWait();
     }
-
 }
